+++ counterparty-core/counterpartycore/lib/bc_data_stream.py
