--- counterparty-core/counterpartycore/lib/kickstart/exceptions.py
@@ -1,6 +1,0 @@
-class SolvingError(Exception):
-    pass
-
-
-class SerializationError(Exception):
-    """Thrown when there's a problem deserializing or serializing"""
