+++ counterparty-wallet/counterpartywallet/messages.py
@@ -8,8 +8,7 @@ from math import ceil
 import bitcoin as bitcoinlib
 import dateutil.parser
 from counterpartycore.lib import config, exceptions, script, transaction
-from counterpartycore.lib.kickstart.utils import ib2h
-from counterpartycore.lib.util import BET_TYPE_ID, dhash
+from counterpartycore.lib.util import BET_TYPE_ID, dhash, ib2h
 
 from counterpartywallet import util, wallet
